Add option to change width and length of hands

+++ wear/src/main/java/com/vlad1m1r/watchface/components/hands/CircleData.kt
@@ -6,7 +6,7 @@ data class CircleData(
     @ColorInt val color: Int,
     @ColorInt val colorAmbient: Int,
     @ColorInt  val shadowColor: Int,
-    val shadowRadius: Float,
-    val width: Float,
-    val radius: Float
+    val shadowRadius: Int,
+    val width: Int,
+    val radius: Int
 )
